Use flat positional view selector on ios/mac
Previously was trying to still use sections that mapped to the adapter's sections but then added an extra item for a header and footer in each section, and then calculated the actual section/index.

This is inconsistent with how it's done for android and windows, where it's a flat list and natively just a single section.  This changes it to be consistent everywhere so it's easier to understand and debug, as there really was little to no advantage to doing it differently here.