Android - no stable item ids
Was returning 'position' for the item id which is potentially wrong if things aren't 'stable' (so if we add / remove items, etc).  According to the docs it should return `NoId` in this case, and we should set `HasStableIds` to `false` to indicate the Id should not be used.

Also, the cached reuse id mapping to int values was being cleared out when the data was invalidated, which means potentially the reuse id's mapped int value could be changing for a given reuse id, causing old and incorrect values to be returned and maybe recycling with a wrong / old template.

Finally, some of the built in item view types have values in the < 100 range, so we'll start our count returning values > 100 to avoid conflicts here.

This should fix some android weirdness that was a result of the incorrect assumptions.