Hide Global/Header Footer when adapter is empty
As described in #31 when the adapter is empty, the global header/footer (which are displayed in cells) overlap the empty view.

While the desired behaviour in some cases might be to make the emptyview fill the remaining space that the global header/footer do not occupy, the problem is the header/footer are implemented as cells in the list (so they scroll with the content of the list - if you didn't want them to scroll, you'd put them outside of the list anyway), so it would be challenging to make the empty view a cell which fills the remaining space.

This change instead hides the global header/footer if the adapter is empty, so that the empty view fills the entire space.  If you really want the same global header/footer to appear when the adapter is empty, you can simply add them also to your empty view template (eg: use a grid with `RowDefinitions="Auto,*Auto"` where your header and footer are rows 0 and 2 and the middle view fills the remaining space.