Fix logic for 'at least one' item
We only show global header/footer if there's at least one item to display.  The header was fine already, but we weren't considering other cases where there were items found in order to decide to display the footer.

So if you had no header visible, but wanted a footer, it wouldn't show up.